Properly handling potion queries in the who-drops function
Required rethinking a little bit about how the fuzzyname function works,
including abandoning the idea of putting lowercase names in the lookup
tables -- that's all around a bad idea!

Perhaps introduces a new bug in the fuzzyname (tying it to items versus being
a little more generic), but we'll cross that bridge if we need to.