Updated Chicken Drop Table
I got a complaint about feather scarcity and looked into the Chicken's feather drop rate. After checking with the 2009 RS Wiki Revision, I noticed the inconsistencies with the current drop table.

Currently, Chickens drop feathers 66% of the time: 33% to drop 1-6, and 33% to drop 15.

Officially, Chickens should drop either 5, 10, or 15 feathers, according to the 2009 RS Wiki Revision.

I've taken the weights from the OSRS Wiki as a reference, so the drop rate for feathers has been changed to 75%: 25% for 5, 25% for 10, 25% for 15, and 25% for none.